requirements.txt
# local package
-e .

# external requirements 
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
ipykernel
pandas
tqdm
matplotlib
seaborn
scipy
scikit-learn
tslearn
kaggle
statsmodels
xgboost
catboost
lightgbm
pyyaml